Arrogate agricultural management system project
The “Argo mate” website’s main goal is to create a dashboard that aids in discovering the most recent information on plant diseases, agricultural industry news, fertilizers to prevent those diseases, information and contact details of sellers and buyers who provide treatments, the locations of showrooms, and other things. The system is being updated and informed by the suppliers of plant diseases. We have information from numerous international suppliers of agricultural products. All are affordable to everyone and are at a fair price. We want to provide each user with the greatest experience.